+++ src/apollo/datastore/utils/admin/SessionsPermissions.java
@@ -8,13 +8,13 @@ public enum SessionsPermissions {
     ALL_PERMISSIONS            ((1 << 4) - 1),
     DEFAULT_PERMISSIONS        (0);
 
-    private final int code;
+    private final long code;
 
-    private SessionsPermissions(int code) {
+    private SessionsPermissions(long code) {
         this.code = code;
     }
 
-    public int getCode() {
+    public long getCode() {
         return this.code;
     }
