+++ src/apollo/datastore/utils/user/SessionLogPermissions.java
@@ -8,13 +8,13 @@ public enum SessionLogPermissions {
     ALL_PERMISSIONS               ((1 << 4) - 1),
     DEFAULT_PERMISSIONS           (0);
 
-    private final int code;
+    private final long code;
 
-    private SessionLogPermissions(int code) {
+    private SessionLogPermissions(long code) {
         this.code = code;
     }
 
-    public int getCode() {
+    public long getCode() {
         return this.code;
     }
